Create dirs and rename namespaces
