+++ Data/Ef/Configurations/ContractConfiguration.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Timesheets.Models;
-
+using Timesheets.Models.Entities;
 
 namespace Timesheets.Data.Ef.Configurations
 {
