+++ Data/Interfaces/IServiceRepo.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Timesheets.Models;
 using System.Threading.Tasks;
+using Timesheets.Models.Entities;
 
 namespace Timesheets.Data.Interfaces
 {
