+++ Infrastructure/Validation/IUserValidationService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Timesheets.Models;
+using Timesheets.Models.Entities;
 
 namespace Timesheets.Infrastructure.Validation
 {
