+++ Models/Dto/EmployeeCreateRequest.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Timesheets.Models.Entities;
 
 namespace Timesheets.Models.Dto
 {
