+++ Models/Enity.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Timesheets.Models
+{
+    public class Enity
+    {
+           public Guid Id { get; set; }
+    }
+}
