+++ Models/Entities/Client.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Timesheets.Models
+namespace Timesheets.Models.Entities
 {
     /// <summary>
     /// Информация о владельце контракта
