+++ Models/Entities/Employee.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Timesheets.Models
+namespace Timesheets.Models.Entities
 {
 
     /// <summary>
