+++ Services/Managers/Interfaces/IUserManager.cs
@@ -4,8 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Timesheets.Models;
 using Timesheets.Models.Dto;
+using Timesheets.Models.Entities;
 
-namespace Timesheets.Services.Interfaces
+namespace Timesheets.Services.Managers.Interfaces
 {
   
     public interface IUserManager
