+++ Services/ValueObject.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Timesheets.Services
+{
+    public class ValueObject
+    {
+    }
+}
