+++ Startup.cs
@@ -1,28 +1,9 @@
-using System;
-using System.Reflection;
-using System.IO;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Microsoft.OpenApi.Models;
-using Timesheets.Data.Implementetion;
-using Timesheets.Data.Interfaces;
-using Timesheets.Services.Implementetion;
-using Timesheets.Services.Interfaces;
-using Timesheets.Data;
-using Microsoft.EntityFrameworkCore;
 using Timesheets.Infrastructure.Extensions;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using System.Text;
-using Timesheets.Models.Dto.Authentication;
 using FluentValidation.AspNetCore;
 
 namespace Timesheets
